pyproject.toml
[tool.poetry]
name = "redditsuite"
version = "0.1.0"
description = ""
authors = ["datavorous <bhattacharjeesagnik910@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
googlesearch-python = "^1.2.5"
requests = "^2.32.3"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
